backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.5
langchain-core==1.0.4
langchain-nvidia-ai-endpoints==0.3.19
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
MarkupSafe==3.0.3
mongoengine==0.29.1
multidict==6.7.0
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.3
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
